Refactor UI helpers into static modules
