+++ packages/api/public/ui-state.js
@@ -1,0 +1,161 @@
+export const state = {
+  shows: [],
+  feeds: [],
+  profiles: [],
+  queries: [],
+  scripts: [],
+  scriptCoachingActions: [],
+  researchPackets: [],
+  modelProfiles: [],
+  promptTemplates: [],
+  scheduledPipelines: [],
+  failedScheduledRuns: [],
+  recentJobs: [],
+  selectedJobId: '',
+  storyCandidates: [],
+  episodePlan: null,
+  episodes: [],
+  selectedScriptId: '',
+  selectedScript: null,
+  selectedRevision: null,
+  selectedRevisions: [],
+  selectedCoverageSummary: null,
+  production: {
+    episode: null,
+    assets: [],
+    jobs: [],
+  },
+  productionPoll: null,
+  jobPoll: null,
+  selectedCandidateIds: [],
+  clusterForm: {
+    angle: '',
+    notes: '',
+    targetFormat: '',
+    targetRuntime: '',
+  },
+  selectedResearchPacketId: '',
+  selectedEpisodeId: '',
+  selectedAssetIds: [],
+  selectedShowSlug: '',
+  selectedProfileId: '',
+  activeSurface: 'workflow',
+  activeSettingsTab: 'shows',
+  showSetupOpen: false,
+  runningActions: {},
+};
+
+export const els = {
+  status: document.querySelector('#status'),
+  errorDetails: document.querySelector('#errorDetails'),
+  errorDetailsBody: document.querySelector('#errorDetailsBody'),
+  refresh: document.querySelector('#refresh'),
+  importLegacy: document.querySelector('#importLegacy'),
+  surfaceTabs: document.querySelectorAll('.surface-tab'),
+  showSelect: document.querySelector('#showSelect'),
+  newShowToggle: document.querySelector('#newShowToggle'),
+  showSetupForm: document.querySelector('#showSetupForm'),
+  showSetupMeta: document.querySelector('#showSetupMeta'),
+  showSetupStatus: document.querySelector('#showSetupStatus'),
+  showName: document.querySelector('#showName'),
+  showSlug: document.querySelector('#showSlug'),
+  showHostVoice: document.querySelector('#showHostVoice'),
+  showDescription: document.querySelector('#showDescription'),
+  showToneNotes: document.querySelector('#showToneNotes'),
+  showScriptNotes: document.querySelector('#showScriptNotes'),
+  showFeedTitle: document.querySelector('#showFeedTitle'),
+  showFeedUrl: document.querySelector('#showFeedUrl'),
+  showPublicBase: document.querySelector('#showPublicBase'),
+  showOutputPath: document.querySelector('#showOutputPath'),
+  showSourceQuery: document.querySelector('#showSourceQuery'),
+  showPublishingMode: document.querySelector('#showPublishingMode'),
+  showModelProvider: document.querySelector('#showModelProvider'),
+  showModelName: document.querySelector('#showModelName'),
+  showReasoningEffort: document.querySelector('#showReasoningEffort'),
+  cancelShowSetup: document.querySelector('#cancelShowSetup'),
+  settingsMeta: document.querySelector('#settingsMeta'),
+  settingsTabs: document.querySelectorAll('.settings-tab'),
+  settingsShows: document.querySelector('#settingsShows'),
+  settingsSources: document.querySelector('#settingsSources'),
+  settingsModels: document.querySelector('#settingsModels'),
+  settingsPrompts: document.querySelector('#settingsPrompts'),
+  settingsPublishing: document.querySelector('#settingsPublishing'),
+  settingsSchedules: document.querySelector('#settingsSchedules'),
+  pipelineMeta: document.querySelector('#pipelineMeta'),
+  workflowContext: document.querySelector('#workflowContext'),
+  nextActionPanel: document.querySelector('#nextActionPanel'),
+  pipelineStages: document.querySelector('#pipelineStages'),
+  pipelineDebug: document.querySelector('#pipelineDebug'),
+  profileList: document.querySelector('#profileList'),
+  profileForm: document.querySelector('#profileForm'),
+  profileTitle: document.querySelector('#profileTitle'),
+  profileMeta: document.querySelector('#profileMeta'),
+  profileEnabled: document.querySelector('#profileEnabled'),
+  profileName: document.querySelector('#profileName'),
+  profileSlug: document.querySelector('#profileSlug'),
+  profileType: document.querySelector('#profileType'),
+  profileWeight: document.querySelector('#profileWeight'),
+  profileFreshness: document.querySelector('#profileFreshness'),
+  profileIncludeDomains: document.querySelector('#profileIncludeDomains'),
+  profileExcludeDomains: document.querySelector('#profileExcludeDomains'),
+  ingestProfile: document.querySelector('#ingestProfile'),
+  manualForm: document.querySelector('#manualForm'),
+  manualUrl: document.querySelector('#manualUrl'),
+  manualTitle: document.querySelector('#manualTitle'),
+  manualSourceName: document.querySelector('#manualSourceName'),
+  manualSummary: document.querySelector('#manualSummary'),
+  manualResult: document.querySelector('#manualResult'),
+  candidateMeta: document.querySelector('#candidateMeta'),
+  candidateClusterForm: document.querySelector('#candidateClusterForm'),
+  selectionCount: document.querySelector('#selectionCount'),
+  selectedCandidateSummary: document.querySelector('#selectedCandidateSummary'),
+  clearCandidateSelection: document.querySelector('#clearCandidateSelection'),
+  selectionWarnings: document.querySelector('#selectionWarnings'),
+  clusterAngle: document.querySelector('#clusterAngle'),
+  clusterNotes: document.querySelector('#clusterNotes'),
+  clusterFormat: document.querySelector('#clusterFormat'),
+  clusterRuntime: document.querySelector('#clusterRuntime'),
+  requestEpisodePlan: document.querySelector('#requestEpisodePlan'),
+  launchClusterBrief: document.querySelector('#launchClusterBrief'),
+  episodePlanResult: document.querySelector('#episodePlanResult'),
+  candidateList: document.querySelector('#candidateList'),
+  researchBriefMeta: document.querySelector('#researchBriefMeta'),
+  researchBriefList: document.querySelector('#researchBriefList'),
+  schedulerMeta: document.querySelector('#schedulerMeta'),
+  schedulerList: document.querySelector('#schedulerList'),
+  failedScheduleRuns: document.querySelector('#failedScheduleRuns'),
+  jobRunsMeta: document.querySelector('#jobRunsMeta'),
+  refreshJobs: document.querySelector('#refreshJobs'),
+  jobRunList: document.querySelector('#jobRunList'),
+  jobRunDetail: document.querySelector('#jobRunDetail'),
+  episodeMeta: document.querySelector('#episodeMeta'),
+  episodeList: document.querySelector('#episodeList'),
+  reviewMeta: document.querySelector('#reviewMeta'),
+  reviewResearch: document.querySelector('#reviewResearch'),
+  reviewScript: document.querySelector('#reviewScript'),
+  reviewProduction: document.querySelector('#reviewProduction'),
+  publishChecklist: document.querySelector('#publishChecklist'),
+  modelMeta: document.querySelector('#modelMeta'),
+  modelProfileList: document.querySelector('#modelProfileList'),
+  queriesPanel: document.querySelector('#queriesPanel'),
+  queryCount: document.querySelector('#queryCount'),
+  newQueryForm: document.querySelector('#newQueryForm'),
+  newQueryText: document.querySelector('#newQueryText'),
+  queryList: document.querySelector('#queryList'),
+  scriptMeta: document.querySelector('#scriptMeta'),
+  scriptGenerateForm: document.querySelector('#scriptGenerateForm'),
+  scriptResearchPacketId: document.querySelector('#scriptResearchPacketId'),
+  scriptFormat: document.querySelector('#scriptFormat'),
+  scriptList: document.querySelector('#scriptList'),
+  scriptEditForm: document.querySelector('#scriptEditForm'),
+  scriptTitle: document.querySelector('#scriptTitle'),
+  scriptBody: document.querySelector('#scriptBody'),
+  scriptCoachingActions: document.querySelector('#scriptCoachingActions'),
+  approveScript: document.querySelector('#approveScript'),
+  productionPanel: document.querySelector('#productionPanel'),
+  productionMeta: document.querySelector('#productionMeta'),
+  generateAudioPreview: document.querySelector('#generateAudioPreview'),
+  generateCoverArt: document.querySelector('#generateCoverArt'),
+  productionJobs: document.querySelector('#productionJobs'),
+  productionAssets: document.querySelector('#productionAssets'),
+};
