+++ packages/api/public/ui.js
@@ -1,238 +1,28 @@
-const state = {
-  shows: [],
-  feeds: [],
-  profiles: [],
-  queries: [],
-  scripts: [],
-  scriptCoachingActions: [],
-  researchPackets: [],
-  modelProfiles: [],
-  promptTemplates: [],
-  scheduledPipelines: [],
-  failedScheduledRuns: [],
-  recentJobs: [],
-  selectedJobId: '',
-  storyCandidates: [],
-  episodePlan: null,
-  episodes: [],
-  selectedScriptId: '',
-  selectedScript: null,
-  selectedRevision: null,
-  selectedRevisions: [],
-  selectedCoverageSummary: null,
-  production: {
-    episode: null,
-    assets: [],
-    jobs: [],
-  },
-  productionPoll: null,
-  jobPoll: null,
-  selectedCandidateIds: [],
-  clusterForm: {
-    angle: '',
-    notes: '',
-    targetFormat: '',
-    targetRuntime: '',
-  },
-  selectedResearchPacketId: '',
-  selectedEpisodeId: '',
-  selectedAssetIds: [],
-  selectedShowSlug: '',
-  selectedProfileId: '',
-  activeSurface: 'workflow',
-  activeSettingsTab: 'shows',
-  showSetupOpen: false,
-  runningActions: {},
-};
-
-const els = {
-  status: document.querySelector('#status'),
-  errorDetails: document.querySelector('#errorDetails'),
-  errorDetailsBody: document.querySelector('#errorDetailsBody'),
-  refresh: document.querySelector('#refresh'),
-  importLegacy: document.querySelector('#importLegacy'),
-  surfaceTabs: document.querySelectorAll('.surface-tab'),
-  showSelect: document.querySelector('#showSelect'),
-  newShowToggle: document.querySelector('#newShowToggle'),
-  showSetupForm: document.querySelector('#showSetupForm'),
-  showSetupMeta: document.querySelector('#showSetupMeta'),
-  showSetupStatus: document.querySelector('#showSetupStatus'),
-  showName: document.querySelector('#showName'),
-  showSlug: document.querySelector('#showSlug'),
-  showHostVoice: document.querySelector('#showHostVoice'),
-  showDescription: document.querySelector('#showDescription'),
-  showToneNotes: document.querySelector('#showToneNotes'),
-  showScriptNotes: document.querySelector('#showScriptNotes'),
-  showFeedTitle: document.querySelector('#showFeedTitle'),
-  showFeedUrl: document.querySelector('#showFeedUrl'),
-  showPublicBase: document.querySelector('#showPublicBase'),
-  showOutputPath: document.querySelector('#showOutputPath'),
-  showSourceQuery: document.querySelector('#showSourceQuery'),
-  showPublishingMode: document.querySelector('#showPublishingMode'),
-  showModelProvider: document.querySelector('#showModelProvider'),
-  showModelName: document.querySelector('#showModelName'),
-  showReasoningEffort: document.querySelector('#showReasoningEffort'),
-  cancelShowSetup: document.querySelector('#cancelShowSetup'),
-  settingsMeta: document.querySelector('#settingsMeta'),
-  settingsTabs: document.querySelectorAll('.settings-tab'),
-  settingsShows: document.querySelector('#settingsShows'),
-  settingsSources: document.querySelector('#settingsSources'),
-  settingsModels: document.querySelector('#settingsModels'),
-  settingsPrompts: document.querySelector('#settingsPrompts'),
-  settingsPublishing: document.querySelector('#settingsPublishing'),
-  settingsSchedules: document.querySelector('#settingsSchedules'),
-  pipelineMeta: document.querySelector('#pipelineMeta'),
-  workflowContext: document.querySelector('#workflowContext'),
-  nextActionPanel: document.querySelector('#nextActionPanel'),
-  pipelineStages: document.querySelector('#pipelineStages'),
-  pipelineDebug: document.querySelector('#pipelineDebug'),
-  profileList: document.querySelector('#profileList'),
-  profileForm: document.querySelector('#profileForm'),
-  profileTitle: document.querySelector('#profileTitle'),
-  profileMeta: document.querySelector('#profileMeta'),
-  profileEnabled: document.querySelector('#profileEnabled'),
-  profileName: document.querySelector('#profileName'),
-  profileSlug: document.querySelector('#profileSlug'),
-  profileType: document.querySelector('#profileType'),
-  profileWeight: document.querySelector('#profileWeight'),
-  profileFreshness: document.querySelector('#profileFreshness'),
-  profileIncludeDomains: document.querySelector('#profileIncludeDomains'),
-  profileExcludeDomains: document.querySelector('#profileExcludeDomains'),
-  ingestProfile: document.querySelector('#ingestProfile'),
-  manualForm: document.querySelector('#manualForm'),
-  manualUrl: document.querySelector('#manualUrl'),
-  manualTitle: document.querySelector('#manualTitle'),
-  manualSourceName: document.querySelector('#manualSourceName'),
-  manualSummary: document.querySelector('#manualSummary'),
-  manualResult: document.querySelector('#manualResult'),
-  candidateMeta: document.querySelector('#candidateMeta'),
-  candidateClusterForm: document.querySelector('#candidateClusterForm'),
-  selectionCount: document.querySelector('#selectionCount'),
-  selectedCandidateSummary: document.querySelector('#selectedCandidateSummary'),
-  clearCandidateSelection: document.querySelector('#clearCandidateSelection'),
-  selectionWarnings: document.querySelector('#selectionWarnings'),
-  clusterAngle: document.querySelector('#clusterAngle'),
-  clusterNotes: document.querySelector('#clusterNotes'),
-  clusterFormat: document.querySelector('#clusterFormat'),
-  clusterRuntime: document.querySelector('#clusterRuntime'),
-  requestEpisodePlan: document.querySelector('#requestEpisodePlan'),
-  launchClusterBrief: document.querySelector('#launchClusterBrief'),
-  episodePlanResult: document.querySelector('#episodePlanResult'),
-  candidateList: document.querySelector('#candidateList'),
-  researchBriefMeta: document.querySelector('#researchBriefMeta'),
-  researchBriefList: document.querySelector('#researchBriefList'),
-  schedulerMeta: document.querySelector('#schedulerMeta'),
-  schedulerList: document.querySelector('#schedulerList'),
-  failedScheduleRuns: document.querySelector('#failedScheduleRuns'),
-  jobRunsMeta: document.querySelector('#jobRunsMeta'),
-  refreshJobs: document.querySelector('#refreshJobs'),
-  jobRunList: document.querySelector('#jobRunList'),
-  jobRunDetail: document.querySelector('#jobRunDetail'),
-  episodeMeta: document.querySelector('#episodeMeta'),
-  episodeList: document.querySelector('#episodeList'),
-  reviewMeta: document.querySelector('#reviewMeta'),
-  reviewResearch: document.querySelector('#reviewResearch'),
-  reviewScript: document.querySelector('#reviewScript'),
-  reviewProduction: document.querySelector('#reviewProduction'),
-  publishChecklist: document.querySelector('#publishChecklist'),
-  modelMeta: document.querySelector('#modelMeta'),
-  modelProfileList: document.querySelector('#modelProfileList'),
-  queriesPanel: document.querySelector('#queriesPanel'),
-  queryCount: document.querySelector('#queryCount'),
-  newQueryForm: document.querySelector('#newQueryForm'),
-  newQueryText: document.querySelector('#newQueryText'),
-  queryList: document.querySelector('#queryList'),
-  scriptMeta: document.querySelector('#scriptMeta'),
-  scriptGenerateForm: document.querySelector('#scriptGenerateForm'),
-  scriptResearchPacketId: document.querySelector('#scriptResearchPacketId'),
-  scriptFormat: document.querySelector('#scriptFormat'),
-  scriptList: document.querySelector('#scriptList'),
-  scriptEditForm: document.querySelector('#scriptEditForm'),
-  scriptTitle: document.querySelector('#scriptTitle'),
-  scriptBody: document.querySelector('#scriptBody'),
-  scriptCoachingActions: document.querySelector('#scriptCoachingActions'),
-  approveScript: document.querySelector('#approveScript'),
-  productionPanel: document.querySelector('#productionPanel'),
-  productionMeta: document.querySelector('#productionMeta'),
-  generateAudioPreview: document.querySelector('#generateAudioPreview'),
-  generateCoverArt: document.querySelector('#generateCoverArt'),
-  productionJobs: document.querySelector('#productionJobs'),
-  productionAssets: document.querySelector('#productionAssets'),
-};
-
-const MODEL_ROLE_LABELS = {
-  episode_planner: {
-    title: 'Episode planner',
-    description: 'Drafts advisory story briefs and episode plans before research begins.',
-  },
-  candidate_scorer: {
-    title: 'Candidate scorer',
-    description: 'Ranks possible stories for editorial fit, significance, novelty, source quality, and urgency.',
-  },
-  source_summarizer: {
-    title: 'Source summarizer',
-    description: 'Summarizes fetched source snapshots without adding unsupported claims.',
-  },
-  claim_extractor: {
-    title: 'Claim extractor',
-    description: 'Turns source material into attributed claims for the research brief.',
-  },
-  research_synthesizer: {
-    title: 'Research synthesizer',
-    description: 'Builds the evidence-first research brief and preserves uncertainty.',
-  },
-  script_writer: {
-    title: 'Script writer',
-    description: 'Drafts the episode script from approved research brief material.',
-  },
-  script_editor: {
-    title: 'Script editor',
-    description: 'Revises drafts while preserving citation and provenance metadata.',
-  },
-  integrity_reviewer: {
-    title: 'Integrity reviewer',
-    description: 'Checks script drafts against source evidence before production.',
-  },
-  metadata_writer: {
-    title: 'Metadata writer',
-    description: 'Creates episode titles, summaries, and feed metadata.',
-  },
-  cover_prompt_writer: {
-    title: 'Cover prompt writer',
-    description: 'Writes art direction for cover image generation.',
-  },
-};
-
-const SETTINGS_SECTIONS = {
-  shows: 'Shows & feeds',
-  sources: 'Story sources/search recipes',
-  models: 'Model roles',
-  prompts: 'Prompt templates',
-  publishing: 'Publishing/storage',
-  schedules: 'Scheduled pipelines',
-};
-
-const SURFACES = new Set(['workflow', 'settings', 'debug']);
-
-class ApiRequestError extends Error {
-  constructor(message, debugDetails) {
-    super(message);
-    this.name = 'ApiRequestError';
-    this.debugDetails = debugDetails;
-  }
-}
-
-function debugText(value) {
-  if (!value) {
-    return '';
-  }
-
-  if (typeof value === 'string') {
-    return value;
-  }
-
-  return JSON.stringify(value, null, 2);
-}
+import { api, ApiRequestError, debugText } from './ui-api.js';
+import { SETTINGS_SECTIONS, SURFACES } from './ui-constants.js';
+import { els, state } from './ui-state.js';
+import {
+  applySourceControlState,
+  applySourceControlStateToForms,
+  asArray,
+  asObject,
+  castToLines,
+  formatRole,
+  linesToCast,
+  linesToList,
+  listToLines,
+  maybeNull,
+  optionalNumber,
+  outputPathForFeed,
+  publicAssetBaseForFeed,
+  readOnboardingSetting,
+  readPublishingMode,
+  roleInfo,
+  safeVisiblePath,
+  sanitizedDebug,
+  slugify,
+  sourceControlsSupported,
+} from './ui-formatters.js';
 
 function setStatus(message, debugDetails = '') {
   els.status.textContent = message;
@@ -250,268 +40,6 @@ function reportError(error, fallback = 'Something went wrong. Open technical det
   const message = error instanceof Error && error.message ? error.message : fallback;
   setStatus(fallback, message);
   return fallback;
-}
-
-function friendlyApiMessage(body, status) {
-  const code = typeof body?.code === 'string' ? body.code : '';
-  const raw = typeof body?.error === 'string' ? body.error : '';
-
-  const messages = {
-    VALIDATION_ERROR: 'Please check the form values and try again.',
-    CONFIG_FILE_NOT_FOUND: 'The requested config file could not be found.',
-    SOURCE_PROFILE_NOT_FOUND: 'That story source could not be found. Refresh and try again.',
-    SOURCE_PROFILE_SHOW_MISMATCH: 'That story source belongs to a different show.',
-    SOURCE_URL_REQUIRED: 'Choose a candidate story with a URL, or add an extra source URL before creating a research brief.',
-    STORY_CANDIDATE_NOT_FOUND: 'That candidate story could not be found. Refresh and try again.',
-    STORY_CANDIDATE_IGNORED: 'Ignored candidate stories cannot be used for research briefs.',
-    CANDIDATE_SHOW_MISMATCH: 'All selected candidate stories must belong to the same show.',
-    EPISODE_PLANNER_RUNTIME_REQUIRED: 'AI episode planning is unavailable because the local LLM runtime is not configured.',
-    EPISODE_PLANNER_MODEL_PROFILE_REQUIRED: 'Configure the episode planner AI role before requesting a plan.',
-    EPISODE_PLAN_MODEL_OUTPUT_INVALID: 'The AI episode planner returned an unreadable plan. No research or approval state changed.',
-    EPISODE_PLAN_MODEL_FAILED: 'The AI episode planner failed. No research or approval state changed.',
-    RESEARCH_PACKET_NOT_FOUND: 'That research brief could not be found. Check the ID and try again.',
-    RESEARCH_PACKET_OR_WARNING_NOT_FOUND: 'That research brief or warning could not be found.',
-    RESEARCH_APPROVAL_BLOCKED: 'Research approval is blocked until the brief is ready and warnings have override reasons.',
-    SCHEDULED_PIPELINE_NOT_FOUND: 'That scheduled pipeline could not be found. Refresh and try again.',
-    SCHEDULED_RUN_NOT_FOUND: 'That scheduled run could not be found. Refresh and try again.',
-    DUPLICATE_SHOW_SLUG: 'Another show already uses that slug. Choose a unique show slug.',
-    DUPLICATE_SOURCE: 'A story source or search query with that unique key already exists.',
-    DUPLICATE_SLUG: 'A record with that slug already exists in this show.',
-    FEED_NOT_FOUND: 'That feed could not be found. Refresh and try again.',
-    MODEL_PROFILE_NOT_FOUND: 'That AI role setting could not be found. Refresh and try again.',
-    INVALID_CRON: 'The cron schedule is not valid. Check the cadence and try again.',
-    SCHEDULED_RUN_NOT_FAILED: 'Only failed scheduled runs can be retried.',
-    PUBLISH_BLOCKED: 'Publishing is blocked until the checklist items are complete.',
-    PUBLISH_APPROVAL_BLOCKED: 'Publish approval is blocked until the checklist items are complete.',
-    SCRIPT_COACHING_RUNTIME_UNAVAILABLE: 'AI script coaching is unavailable because the local LLM runtime is not configured.',
-    SCRIPT_COACHING_MODEL_PROFILE_REQUIRED: 'Configure the script editor AI role before using coaching actions.',
-    MALFORMED_MODEL_OUTPUT: 'The AI returned an unreadable script revision. No approval or revision state changed.',
-    MODEL_INVOCATION_FAILED: 'The AI script editor failed. No approval or revision state changed.',
-  };
-
-  if (messages[code]) {
-    return messages[code];
-  }
-
-  if (status === 404) {
-    return 'That record could not be found. Refresh and try again.';
-  }
-
-  if (status === 409) {
-    return raw || 'This action is blocked by the current review or publishing state.';
-  }
-
-  if (status >= 500) {
-    return 'The local API hit a server error. Open technical details for the response.';
-  }
-
-  return raw || `Request failed with status ${status}.`;
-}
-
-function linesToList(value) {
-  return value
-    .split(/[\n,]/)
-    .map((item) => item.trim())
-    .filter(Boolean);
-}
-
-function listToLines(value) {
-  return (value || []).join('\n');
-}
-
-function sourceControlsSupported(type) {
-  return type === 'brave' || type === 'rss';
-}
-
-function sourceControlHelp(type) {
-  if (type === 'brave') {
-    return 'Freshness is sent to Brave. Domain filters are enforced after results return.';
-  }
-
-  if (type === 'rss') {
-    return 'Domain filters are enforced on item URLs. Freshness is checked when feed items include published dates.';
-  }
-
-  return 'Freshness and domain filters are not applied for manual or local JSON sources.';
-}
-
-function applySourceControlState(root, type) {
-  const supported = sourceControlsSupported(type);
-  const fields = [
-    root.querySelector('[name="freshness"], #profileFreshness'),
-    root.querySelector('[name="includeDomains"], #profileIncludeDomains'),
-    root.querySelector('[name="excludeDomains"], #profileExcludeDomains'),
-  ].filter(Boolean);
-
-  for (const field of fields) {
-    field.disabled = !supported;
-    if (!supported) {
-      field.value = '';
-    }
-  }
-
-  const help = root.querySelector('[data-source-control-help]');
-  if (help) {
-    help.textContent = sourceControlHelp(type);
-  }
-}
-
-function applySourceControlStateToForms(root, type) {
-  for (const form of root.querySelectorAll('form')) {
-    applySourceControlState(form, type);
-  }
-}
-
-function asObject(value) {
-  return value && typeof value === 'object' && !Array.isArray(value) ? value : {};
-}
-
-function asArray(value) {
-  return Array.isArray(value) ? value : [];
-}
-
-function maybeNull(value) {
-  const trimmed = String(value || '').trim();
-  return trimmed ? trimmed : null;
-}
-
-function optionalNumber(value) {
-  const trimmed = String(value || '').trim();
-  return trimmed ? Number(trimmed) : null;
-}
-
-function slugify(value) {
-  return value
-    .toLowerCase()
-    .replace(/[^a-z0-9]+/g, '-')
-    .replace(/^-|-$/g, '')
-    .replace(/-{2,}/g, '-');
-}
-
-function roleInfo(role) {
-  return MODEL_ROLE_LABELS[role] || {
-    title: role.replaceAll('_', ' '),
-    description: 'Configured AI role for this show.',
-  };
-}
-
-function formatRole(role) {
-  return roleInfo(role).title;
-}
-
-function readOnboardingSetting(show, key, fallback = '') {
-  const onboarding = asObject(show?.settings?.onboarding);
-  return typeof onboarding[key] === 'string' ? onboarding[key] : fallback;
-}
-
-function readPublishingMode(show) {
-  return readOnboardingSetting(show, 'publishingMode', 'approval-gated');
-}
-
-function outputPathForFeed(feed) {
-  const metadata = asObject(feed.metadata);
-  const storageConfig = asObject(feed.storageConfig);
-  const value = metadata.outputPath || storageConfig.outputPath || feed.rssFeedPath || '';
-  return typeof value === 'string' ? safeVisiblePath(value) : '';
-}
-
-function publicAssetBaseForFeed(feed) {
-  const metadata = asObject(feed.metadata);
-  return typeof metadata.publicAssetBaseUrl === 'string' ? metadata.publicAssetBaseUrl : feed.publicBaseUrl || '';
-}
-
-function safeVisiblePath(value) {
-  const text = String(value || '').trim();
-
-  if (!text) {
-    return '';
-  }
-
-  if (text.startsWith('/') || text.startsWith('~') || /^[A-Za-z]:[\\/]/.test(text)) {
-    return '';
-  }
-
-  return text;
-}
-
-function sanitizedDebug(value) {
-  if (!value || typeof value !== 'object') {
-    return value;
-  }
-
-  if (Array.isArray(value)) {
-    return value.map(sanitizedDebug);
-  }
-
-  const result = {};
-  for (const [key, item] of Object.entries(value)) {
-    if (/secret|token|password|credential|api.?key|local.?path|private.?key/i.test(key)) {
-      result[key] = '[hidden]';
-    } else if (typeof item === 'string' && (item.startsWith('/') || item.startsWith('~'))) {
-      result[key] = '[hidden local path]';
-    } else {
-      result[key] = sanitizedDebug(item);
-    }
-  }
-
-  return result;
-}
-
-function castToLines(cast) {
-  return asArray(cast)
-    .map((member) => [member.name, member.role || '', member.voice].filter((part) => part !== '').join(' | '))
-    .join('\n');
-}
-
-function linesToCast(value) {
-  return value
-    .split('\n')
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .map((line) => {
-      const [name, roleOrVoice, voice] = line.split('|').map((part) => part.trim());
-      return {
-        name,
-        ...(voice ? { role: roleOrVoice } : {}),
-        voice: voice || roleOrVoice || name,
-      };
-    });
-}
-
-async function api(path, options = {}) {
-  const response = await fetch(path, {
-    ...options,
-    headers: {
-      'content-type': 'application/json',
-      ...(options.headers || {}),
-    },
-  });
-
-  if (response.status === 204) {
-    return undefined;
-  }
-
-  const text = await response.text();
-  let body;
-
-  try {
-    body = text ? JSON.parse(text) : undefined;
-  } catch (error) {
-    throw new ApiRequestError(
-      response.ok ? 'The API returned an unreadable response.' : `Request failed with status ${response.status}.`,
-      text,
-    );
-  }
-
-  if (!response.ok || body.ok === false) {
-    throw new ApiRequestError(friendlyApiMessage(body, response.status), {
-      path,
-      status: response.status,
-      response: body,
-    });
-  }
-
-  return body;
 }
 
 function selectedProfile() {
